split up hpack.c
aws_hpack_context did too many things and was difficult to understand. Split the encoding and decoding parts out into their own structs and files.

haven't fixed up the tests yet
